feat(replay): PR147 production convergence Slice A — Protocol Core
- strict contiguous framing（ReplayPacketParser + ReplayPacketStreamReader）：
  payloadLen==0 合法、0xFFFFFFFF terminator、禁逐 byte resync、trailing garbage FAIL
- version gate（ReplayVersionGate）：closed semantics 仅 11.19/11.18 版本族 AFFIRMED，
  未知版本 raw-preserve；decode context 改用流头权威 clientVersion
- Type10 49B：attachmentParentEntityId / trailingStateRaw（UNKNOWN 语义，删除 is_error）
- Type10MovementMath：1 unit≈1m、speed km/h、forward/reverse、yaw rate（§A4）

+++ java/wotb-core/src/test/java/com/wotb/core/replay/feature/DefaultTeamBattleFeatureExtractorTest.java
@@ -1159,7 +1159,7 @@ class DefaultTeamBattleFeatureExtractorTest {
         return new PositionChangedEvent(
                 sequence, new ReplayTimestamp(time, null), 10,
                 confidence, entityId, 0, 0,
-                x, 0, z, 0, 0, 0, 0, 0, 0, (byte) 0);
+                x, 0, z, 0, 0, 0, 0, 0, 0, 0);
     }
 
     private static PositionChangedEvent positionWithElevation(
