+++ java/wotb-core/src/test/java/com/wotb/core/replay/timeline/TimelineTestFixtures.java
@@ -82,7 +82,7 @@ final class TimelineTestFixtures {
     static PositionChangedEvent position(final int eid, final double battleSec,
                                          final float x, final float z, final float yawRad) {
         return new PositionChangedEvent(seq++, ts(battleSec), 10, DecodeConfidence.EXACT,
-                eid, 0, 0, x, 0f, z, 0f, 0f, 0f, yawRad, 0f, 0f, (byte) 0);
+                eid, 0, 0, x, 0f, z, 0f, 0f, 0f, yawRad, 0f, 0f, 0);
     }
 
     static HealthChangedEvent health(final int eid, final double battleSec,
